refactor: change uint32_t on int16_t for random numbers method

+++ engine_utility.cpp
@@ -4,7 +4,7 @@ namespace dr
 {
   namespace EngineUtility
   {
-    uint32_t getRandomInRange(uint32_t min, uint32_t max)
+    int16_t getRandomInRange(int16_t min, int16_t max)
     {
       std::random_device rd;
       std::mt19937 rn{ rd() };
